+++ NEWS.md
@@ -1,3 +1,6 @@
+* Added tables of the unique machines, useful work products, and machine-useful work product combinations to the "Framework" tab of the App.
+
+
 # PFU-Database 0.0.2 (2020-09-18)
 
 * First version of PFU-Database which contains a working version of the Shiny App.